requirements.txt
urllib3==1.26.6
databricks-connect==13.0.0
pytest
